add pinky model
