+++ hand-gesture-recognition/model/hand_landmarks.png
