+++ hand-gesture-recognition/model/keypoint_classifier/keypoint_27nov/keypoint_classifier_label.csv
@@ -1,4 +1,4 @@
 ﻿Zoom
 Rotate
-Tilt (UD)
 Tilt (LR)
+Pinky
